+++ butterknife-compiler/src/main/java/butterknife/compiler/ButterKnifeProcessor.java
@@ -30,6 +30,7 @@ import butterknife.internal.ListenerClass;
 import butterknife.internal.ListenerMethod;
 import com.google.auto.common.SuperficialValidation;
 import com.google.auto.service.AutoService;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.squareup.javapoet.JavaFile;
 import com.squareup.javapoet.TypeName;
@@ -48,6 +49,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.BitSet;
 import java.util.Deque;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -72,9 +74,10 @@ import javax.lang.model.type.DeclaredType;
 import javax.lang.model.type.TypeKind;
 import javax.lang.model.type.TypeMirror;
 import javax.lang.model.type.TypeVariable;
-import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
 import javax.tools.Diagnostic.Kind;
+import net.ltgt.gradle.incap.IncrementalAnnotationProcessor;
+import net.ltgt.gradle.incap.IncrementalAnnotationProcessorType;
 
 import static butterknife.internal.Constants.NO_RES_ID;
 import static java.util.Objects.requireNonNull;
@@ -85,6 +88,7 @@ import static javax.lang.model.element.Modifier.PRIVATE;
 import static javax.lang.model.element.Modifier.STATIC;
 
 @AutoService(Processor.class)
+@IncrementalAnnotationProcessor(IncrementalAnnotationProcessorType.DYNAMIC)
 @SuppressWarnings("NullAway") // TODO fix all these...
 public final class ButterKnifeProcessor extends AbstractProcessor {
   // TODO remove when http://b.android.com/187527 is released.
@@ -123,7 +127,6 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
 
   private int sdk = 1;
   private boolean debuggable = true;
-  private boolean useLegacyTypes = false;
 
   private final RScanner rScanner = new RScanner();
 
@@ -143,18 +146,34 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
     }
 
     debuggable = !"false".equals(env.getOptions().get(OPTION_DEBUGGABLE));
-    useLegacyTypes = !hasAndroidX(env.getElementUtils());
 
     typeUtils = env.getTypeUtils();
     filer = env.getFiler();
     try {
       trees = Trees.instance(processingEnv);
     } catch (IllegalArgumentException ignored) {
+      try {
+        // Get original ProcessingEnvironment from Gradle-wrapped one or KAPT-wrapped one.
+        for (Field field : processingEnv.getClass().getDeclaredFields()) {
+          if (field.getName().equals("delegate") || field.getName().equals("processingEnv")) {
+            field.setAccessible(true);
+            ProcessingEnvironment javacEnv = (ProcessingEnvironment) field.get(processingEnv);
+            trees = Trees.instance(javacEnv);
+            break;
+          }
+        }
+      } catch (Throwable ignored2) {
+      }
     }
   }
 
   @Override public Set<String> getSupportedOptions() {
-    return ImmutableSet.of(OPTION_SDK_INT, OPTION_DEBUGGABLE);
+    ImmutableSet.Builder<String> builder = ImmutableSet.builder();
+    builder.add(OPTION_SDK_INT, OPTION_DEBUGGABLE);
+    if (trees != null) {
+      builder.add(IncrementalAnnotationProcessorType.ISOLATING.getProcessorOption());
+    }
+    return builder.build();
   }
 
   @Override public Set<String> getSupportedAnnotationTypes() {
@@ -193,7 +212,7 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
       TypeElement typeElement = entry.getKey();
       BindingSet binding = entry.getValue();
 
-      JavaFile javaFile = binding.brewJava(sdk, debuggable, useLegacyTypes);
+      JavaFile javaFile = binding.brewJava(sdk, debuggable);
       try {
         javaFile.writeTo(filer);
       } catch (IOException e) {
@@ -345,6 +364,9 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
       findAndParseListener(env, listener, builderMap, erasedTargetNames);
     }
 
+    Map<TypeElement, ClasspathBindingSet> classpathBindings =
+        findAllSupertypeBindings(builderMap, erasedTargetNames);
+
     // Associate superclass binders with their subclass binders. This is a queue-based tree walk
     // which starts at the roots (superclasses) and walks to the leafs (subclasses).
     Deque<Map.Entry<TypeElement, BindingSet.Builder>> entries =
@@ -356,11 +378,14 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
       TypeElement type = entry.getKey();
       BindingSet.Builder builder = entry.getValue();
 
-      TypeElement parentType = findParentType(type, erasedTargetNames);
+      TypeElement parentType = findParentType(type, erasedTargetNames, classpathBindings.keySet());
       if (parentType == null) {
         bindingMap.put(type, builder.build());
       } else {
-        BindingSet parentBinding = bindingMap.get(parentType);
+        BindingInformationProvider parentBinding = bindingMap.get(parentType);
+        if (parentBinding == null) {
+          parentBinding = classpathBindings.get(parentType);
+        }
         if (parentBinding != null) {
           builder.setParent(parentBinding);
           bindingMap.put(type, builder.build());
@@ -671,12 +696,8 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
     Id resourceId = elementToId(element, BindColor.class, id);
     BindingSet.Builder builder = getOrCreateBindingBuilder(builderMap, enclosingElement);
 
-    FieldResourceBinding.Type colorStateList = useLegacyTypes
-        ? FieldResourceBinding.Type.COLOR_STATE_LIST_LEGACY
-            : FieldResourceBinding.Type.COLOR_STATE_LIST;
-    FieldResourceBinding.Type color = useLegacyTypes
-        ? FieldResourceBinding.Type.COLOR_LEGACY
-        : FieldResourceBinding.Type.COLOR;
+    FieldResourceBinding.Type colorStateList = FieldResourceBinding.Type.COLOR_STATE_LIST;
+    FieldResourceBinding.Type color = FieldResourceBinding.Type.COLOR;
     builder.addResource(new FieldResourceBinding(
         resourceId,
         name,
@@ -781,8 +802,7 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
     Map<Integer, Id> resourceIds = elementToIds(element, BindDrawable.class, new int[] {id, tint});
 
     BindingSet.Builder builder = getOrCreateBindingBuilder(builderMap, enclosingElement);
-    builder.addResource(new FieldDrawableBinding(resourceIds.get(id), name, resourceIds.get(tint),
-        useLegacyTypes));
+    builder.addResource(new FieldDrawableBinding(resourceIds.get(id), name, resourceIds.get(tint)));
 
     erasedTargetNames.add(enclosingElement);
   }
@@ -854,7 +874,7 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
 
     BindingSet.Builder builder = getOrCreateBindingBuilder(builderMap, enclosingElement);
     Id resourceId = elementToId(element, BindFont.class, bindFont.value());
-    builder.addResource(new FieldTypefaceBinding(resourceId, name, style, useLegacyTypes));
+    builder.addResource(new FieldTypefaceBinding(resourceId, name, style));
 
     erasedTargetNames.add(enclosingElement);
   }
@@ -1118,7 +1138,9 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
       TypeVariable typeVariable = (TypeVariable) returnType;
       returnType = typeVariable.getUpperBound();
     }
-    if (!returnType.toString().equals(method.returnType())) {
+    String returnTypeString = returnType.toString();
+    boolean hasReturnValue = !"void".equals(returnTypeString);
+    if (!returnTypeString.equals(method.returnType()) && hasReturnValue) {
       error(element, "@%s methods must have a '%s' return type. (%s.%s)",
           annotationClass.getSimpleName(), method.returnType(),
           enclosingElement.getQualifiedName(), element.getSimpleName());
@@ -1194,7 +1216,8 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
       }
     }
 
-    MethodViewBinding binding = new MethodViewBinding(name, Arrays.asList(parameters), required);
+    MethodViewBinding binding =
+        new MethodViewBinding(name, Arrays.asList(parameters), required, hasReturnValue);
     BindingSet.Builder builder = getOrCreateBindingBuilder(builderMap, enclosingElement);
     Map<Integer, Id> resourceIds = elementToIds(element, annotationClass, ids);
 
@@ -1269,19 +1292,86 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
     return builder;
   }
 
-  /** Finds the parent binder type in the supplied set, if any. */
-  private @Nullable TypeElement findParentType(TypeElement typeElement, Set<TypeElement> parents) {
-    TypeMirror type;
+  /** Finds the parent binder type in the supplied sets, if any. */
+  private @Nullable TypeElement findParentType(
+      TypeElement typeElement, Set<TypeElement> parents, Set<TypeElement> classpathParents) {
     while (true) {
-      type = typeElement.getSuperclass();
-      if (type.getKind() == TypeKind.NONE) {
-        return null;
-      }
-      typeElement = (TypeElement) ((DeclaredType) type).asElement();
-      if (parents.contains(typeElement)) {
+      typeElement = getSuperClass(typeElement);
+      if (typeElement == null || parents.contains(typeElement)
+          || classpathParents.contains(typeElement)) {
         return typeElement;
       }
     }
+  }
+
+  private Map<TypeElement, ClasspathBindingSet> findAllSupertypeBindings(
+      Map<TypeElement, BindingSet.Builder> builderMap, Set<TypeElement> processedInThisRound) {
+    Map<TypeElement, ClasspathBindingSet> classpathBindings = new HashMap<>();
+
+    Set<Class<? extends Annotation>> supportedAnnotations = getSupportedAnnotations();
+    Set<Class<? extends Annotation>> requireViewInConstructor =
+        ImmutableSet.<Class<? extends Annotation>>builder()
+            .addAll(LISTENERS).add(BindView.class).add(BindViews.class).build();
+    supportedAnnotations.removeAll(requireViewInConstructor);
+
+    for (TypeElement typeElement : builderMap.keySet()) {
+      // Make sure to process superclass before subclass. This is because if there is a class that
+      // requires a View in the constructor, all subclasses need it as well.
+      Deque<TypeElement> superClasses = new ArrayDeque<>();
+      TypeElement superClass = getSuperClass(typeElement);
+      while (superClass != null && !processedInThisRound.contains(superClass)
+          && !classpathBindings.containsKey(superClass)) {
+        superClasses.addFirst(superClass);
+        superClass = getSuperClass(superClass);
+      }
+
+      boolean parentHasConstructorWithView = false;
+      while (!superClasses.isEmpty()) {
+        TypeElement superclass = superClasses.removeFirst();
+        ClasspathBindingSet classpathBinding =
+            findBindingInfoForType(superclass, requireViewInConstructor, supportedAnnotations,
+                parentHasConstructorWithView);
+        if (classpathBinding != null) {
+          parentHasConstructorWithView |= classpathBinding.constructorNeedsView();
+          classpathBindings.put(superclass, classpathBinding);
+        }
+      }
+    }
+    return ImmutableMap.copyOf(classpathBindings);
+  }
+
+  private @Nullable ClasspathBindingSet findBindingInfoForType(
+      TypeElement typeElement, Set<Class<? extends Annotation>> requireConstructorWithView,
+      Set<Class<? extends Annotation>> otherAnnotations, boolean needsConstructorWithView) {
+    boolean foundSupportedAnnotation = false;
+    for (Element enclosedElement : typeElement.getEnclosedElements()) {
+      for (Class<? extends Annotation> bindViewAnnotation : requireConstructorWithView) {
+        if (enclosedElement.getAnnotation(bindViewAnnotation) != null) {
+          return new ClasspathBindingSet(true, typeElement);
+        }
+      }
+      for (Class<? extends Annotation> supportedAnnotation : otherAnnotations) {
+        if (enclosedElement.getAnnotation(supportedAnnotation) != null) {
+          if (needsConstructorWithView) {
+            return new ClasspathBindingSet(true, typeElement);
+          }
+          foundSupportedAnnotation = true;
+        }
+      }
+    }
+    if (foundSupportedAnnotation) {
+      return new ClasspathBindingSet(false, typeElement);
+    } else {
+      return null;
+    }
+  }
+
+  private @Nullable TypeElement getSuperClass(TypeElement typeElement) {
+    TypeMirror type = typeElement.getSuperclass();
+    if (type.getKind() == TypeKind.NONE) {
+      return null;
+    }
+    return (TypeElement) ((DeclaredType) type).asElement();
   }
 
   @Override public SourceVersion getSupportedSourceVersion() {
@@ -1359,14 +1449,6 @@ public final class ButterKnifeProcessor extends AbstractProcessor {
       }
     }
     return null;
-  }
-
-  /**
-   * Check for an AndroidX type required by the runtime to determine whether we're in AndroidX or
-   * using legacy support library types.
-   */
-  private static boolean hasAndroidX(Elements elements) {
-    return elements.getTypeElement("androidx.core.content.ContextCompat") != null;
   }
 
   private static class RScanner extends TreeScanner {