+++ butterknife-gradle-plugin/src/main/java/butterknife/plugin/FinalRClassBuilder.kt
@@ -4,14 +4,11 @@ import com.squareup.javapoet.ClassName
 import com.squareup.javapoet.FieldSpec
 import com.squareup.javapoet.JavaFile
 import com.squareup.javapoet.TypeSpec
-import java.util.Locale
-import javax.lang.model.element.Modifier.FINAL
-import javax.lang.model.element.Modifier.PUBLIC
-import javax.lang.model.element.Modifier.STATIC
+import java.util.*
+import javax.lang.model.element.Modifier.*
 
 
-private const val ANNOTATION_PACKAGE = "androidx.annotation"
-private const val ANNOTATION_PACKAGE_LEGACY = "android.support.annotation"
+private const val ANNOTATION_PACKAGE = "android.support.annotation"
 internal val SUPPORTED_TYPES = setOf("anim", "array", "attr", "bool", "color", "dimen",
     "drawable", "id", "integer", "layout", "menu", "plurals", "string", "style", "styleable")
 
@@ -21,8 +18,7 @@ internal val SUPPORTED_TYPES = setOf("anim", "array", "attr", "bool", "color", "
  */
 class FinalRClassBuilder(
   private val packageName: String,
-  private val className: String,
-  private val useLegacyTypes: Boolean
+  private val className: String
 ) {
 
   private var resourceTypes = mutableMapOf<String, TypeSpec.Builder>()
@@ -58,8 +54,7 @@ class FinalRClassBuilder(
   }
 
   private fun getSupportAnnotationClass(type: String): ClassName {
-    val supportPackage = if (useLegacyTypes) ANNOTATION_PACKAGE_LEGACY else ANNOTATION_PACKAGE
-    return ClassName.get(supportPackage, type.capitalize(Locale.US) + "Res")
+    return ClassName.get(ANNOTATION_PACKAGE, type.capitalize(Locale.US) + "Res")
   }
 
   // TODO https://youtrack.jetbrains.com/issue/KT-28933